refactor(export): split the export wire contract out of the rendering engine
MeshWeaver.Blazor referenced MeshWeaver.Markdown.Export in full, so the document
renderer — and with it Markdig, DocumentFormat.OpenXml and HtmlAgilityPack —
sat in the Blazor layer's transitive closure. The dialog never renders anything:
it posts an ExportDocumentRequest and reads the bytes back off the Activity. It
needs the wire shapes, not the engine.

Move those shapes into a new MeshWeaver.Markdown.Export.Contract:
ExportDocumentRequest/Response, DocumentExportOptions, ExportFormat,
ExportFidelity, RenderedDocument, CorporateIdentity, and AddMarkdownExportTypes.
MeshWeaver.Blazor now references only that. Verified against the build output:
MeshWeaver.Markdown.Export.dll, DocumentFormat.OpenXml.dll and HtmlAgilityPack.dll
are gone from MeshWeaver.Blazor/bin.

Namespaces are unchanged, so every `using MeshWeaver.Markdown.Export.Configuration;`
call site — including the three .csx templates — is untouched. The $type
discriminators come from nameof(), so they are keyed off the type NAME, not its
assembly: the React, React Native and portal-next clients, which post the literal
string "ExportDocumentRequest", are unaffected.

ExportDocumentControl deliberately stays in MeshWeaver.Layout. Layout is the
control vocabulary the server and all four renderers agree on, so a UiControl
belongs there by construction; the six types moved here are messages, options and
node content, which do not.

🚨 The contract gets its OWN KernelScriptAssembly registration. That marker is
per-assembly, and the .csx export templates now span both halves. Omitting it
compiles clean and fails only at runtime on the first export — ExportTemplateCompilationTest
is what pins it, and it passes.

Tests: ExportTemplateCompilationTest 4/4, MeshWeaver.Markdown.Export.Test 99/99,
Monolith export/deck/send 36/36. Release + -warnaserror clean across Contract,
engine, Blazor, Portal.Shared, Portal.Distributed, Orleans/Security/Monolith/Export tests.

No What's New entry: pure-internal layering change, no user-visible effect.
