+++ src/MeshWeaver.Blazor/MeshWeaver.Blazor.csproj
@@ -30,7 +30,11 @@
         <ProjectReference Include="..\MeshWeaver.Kernel\MeshWeaver.Kernel.csproj" />
         <ProjectReference Include="..\MeshWeaver.Markdown\MeshWeaver.Markdown.csproj" />
         <ProjectReference Include="..\MeshWeaver.Markdown.Collaboration\MeshWeaver.Markdown.Collaboration.csproj" />
-        <ProjectReference Include="..\MeshWeaver.Markdown.Export\MeshWeaver.Markdown.Export.csproj" />
+        <!-- The export CONTRACT only — never the engine. The export dialog posts a request and
+             reads the rendered bytes back off the Activity; it does not render anything itself,
+             so the renderer (and Markdig / DocumentFormat.OpenXml / HtmlAgilityPack behind it)
+             has no business in the Blazor layer's dependency graph. -->
+        <ProjectReference Include="..\MeshWeaver.Markdown.Export.Contract\MeshWeaver.Markdown.Export.Contract.csproj" />
         <ProjectReference Include="..\MeshWeaver.Mesh.Contract\MeshWeaver.Mesh.Contract.csproj" />
         <ProjectReference Include="..\MeshWeaver.Layout\MeshWeaver.Layout.csproj" />
     </ItemGroup>
