+++ src/MeshWeaver.Markdown.Export.Contract/Messaging/RenderedDocument.cs
